front: move searched op sorting to new file
