+++ front/src/common/Map/Search/consts.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export const MAIN_OP_CH_CODES = ['', '00', 'BV'];
